// src/types/weather.ts
// Generated by https://quicktype.io

export type TWeather = {
	lat: number;
	lon: number;
	timezone: string;
	timezone_offset: number;
	daily: DailyWeather[];
};

export type CurrentWeather = {
	dt: number;
	sunrise: number;
	sunset: number;
	temp: number;
	feels_like: number;
	pressure: number;
	humidity: number;
	dew_point: number;
	uvi: number;
	clouds: number;
	visibility: number;
	wind_speed: number;
	wind_deg: number;
	weather: Weather[];
};

export type Weather = {
	id: number;
	main: Main;
	description: Description;
	icon: Icon;
};

export enum Description {
	BrokenClouds = 'broken clouds',
	ClearSky = 'clear sky',
	FewClouds = 'few clouds',
	LightRain = 'light rain',
	ModerateRain = 'moderate rain',
	OvercastClouds = 'overcast clouds',
	ScatteredClouds = 'scattered clouds',
}

export enum Icon {
	The01D = '01d',
	The02D = '02d',
	The03D = '03d',
	The04D = '04d',
	The04N = '04n',
	The10D = '10d',
	The10N = '10n',
}

export enum Main {
	Clear = 'Clear',
	Clouds = 'Clouds',
	Rain = 'Rain',
}

export type DailyWeather = {
	dt: number;
	sunrise: number;
	sunset: number;
	temp: Temp;
	feels_like: FeelsLike;
	pressure: number;
	humidity: number;
	dew_point: number;
	wind_speed: number;
	wind_deg: number;
	weather: Weather[];
	clouds: number;
	uvi: number;
	rain?: number;
};

export type FeelsLike = {
	day: number;
	night: number;
	eve: number;
	morn: number;
};

export type Temp = {
	day: number;
	min: number;
	max: number;
	night: number;
	eve: number;
	morn: number;
};

export type Hourly = {
	dt: number;
	temp: number;
	feels_like: number;
	pressure: number;
	humidity: number;
	dew_point: number;
	clouds: number;
	wind_speed: number;
	wind_deg: number;
	weather: Weather[];
	rain?: Rain;
};

export type Rain = {
	'1h': number;
};